model_server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Computer vision + detection
ultralytics>=8.0.0
opencv-python-headless>=4.8.0

# Embeddings
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=10.0.0

# Vector search
faiss-cpu>=1.7.4

# Numerics
numpy>=1.24.0
scipy>=1.11.0
